# docs/generated/openapi.rapidapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Naruto Mythos TCG Developer API",
    "description": "A developer-friendly REST API for querying the Naruto Mythos TCG catalogue. Card is a canonical identity; Printing is a collectible identity; Edition is scoped to a Set.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/sets/{set_id}/editions": {
      "get": {
        "tags": [
          "Editions",
          "Collector Catalogue"
        ],
        "summary": "List a Set's Editions",
        "description": "Editions are scoped to exactly one Set. Edition counts distinguish unique canonical Cards from collectible Printings.",
        "operationId": "list_set_editions_v1_sets__set_id__editions_get",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Set Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEditionsResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Set not found.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SET_NOT_FOUND",
                    "message": "Set not found."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/editions/{edition_id}": {
      "get": {
        "tags": [
          "Editions",
          "Collector Catalogue"
        ],
        "summary": "Get an Edition",
        "description": "An Edition is a Set-specific release scope. Collector status is VERIFIED only when a curated checklist is attached.",
        "operationId": "get_edition_v1_editions__edition_id__get",
        "parameters": [
          {
            "name": "edition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Edition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditionDetail"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Edition not found.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "EDITION_NOT_FOUND",
                    "message": "Edition not found."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/editions/{edition_id}/printings": {
      "get": {
        "tags": [
          "Editions",
          "Collector Catalogue"
        ],
        "summary": "List an Edition's verified collector catalogue",
        "description": "Returns curated collector-reference identities when available. This intentionally differs from GET /v1/printings, which returns every persisted Printing including source-only rows.",
        "operationId": "edition_printings_v1_editions__edition_id__printings_get",
        "parameters": [
          {
            "name": "edition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Edition Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPrintingsResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Edition not found.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "EDITION_NOT_FOUND",
                    "message": "Edition not found."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination or sort parameters.",
            "content": {
              "application/json": {
                "examples": {
                  "pagination": {
                    "value": {
                      "error": {
                        "code": "INVALID_PAGINATION",
                        "message": "page must be >= 1."
                      }
                    }
                  },
                  "sort": {
                    "value": {
                      "error": {
                        "code": "INVALID_SORT",
                        "message": "Invalid sort field."
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/rarities": {
      "get": {
        "tags": [
          "Metadata"
        ],
        "summary": "List legacy canonical Card rarity counts",
        "description": "Legacy compatibility endpoint. card_count groups the canonical Card.rarity projection and is not a collectible Printing count. Use /v1/printing-rarities for Printing-level rarity counts.",
        "operationId": "rarities_v1_rarities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RarityCatalogItem"
                  },
                  "type": "array",
                  "title": "Response Rarities V1 Rarities Get"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/printing-rarities": {
      "get": {
        "tags": [
          "Metadata"
        ],
        "summary": "List collectible Printing rarity counts",
        "operationId": "printing_rarities_v1_printing_rarities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrintingRarityCatalogItem"
                  },
                  "type": "array",
                  "title": "Response Printing Rarities V1 Printing Rarities Get"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/keywords": {
      "get": {
        "tags": [
          "Metadata"
        ],
        "summary": "List keyword counts",
        "operationId": "keywords_v1_keywords_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KeywordCatalogItem"
                  },
                  "type": "array",
                  "title": "Response Keywords V1 Keywords Get"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/keywords/{slug}/cards": {
      "get": {
        "tags": [
          "Metadata"
        ],
        "summary": "List a keyword's cards",
        "operationId": "keyword_cards_v1_keywords__slug__cards_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Page number, >= 1.",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number, >= 1."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Items per page, 1 through 100.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Items per page, 1 through 100."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "number, name, rarity, set, release_date.",
              "default": "number",
              "title": "Sort"
            },
            "description": "number, name, rarity, set, release_date."
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "asc or desc; nulls last.",
              "default": "asc",
              "title": "Order"
            },
            "description": "asc or desc; nulls last."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCardsResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination or sort parameters.",
            "content": {
              "application/json": {
                "examples": {
                  "pagination": {
                    "value": {
                      "error": {
                        "code": "INVALID_PAGINATION",
                        "message": "page must be >= 1."
                      }
                    }
                  },
                  "sort": {
                    "value": {
                      "error": {
                        "code": "INVALID_SORT",
                        "message": "Invalid sort field."
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Keyword not found.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "KEYWORD_NOT_FOUND",
                    "message": "Keyword not found."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EditionDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "canonical_card_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Canonical Card Count"
          },
          "collectible_printing_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Collectible Printing Count"
          },
          "collector_reference_status": {
            "type": "string",
            "pattern": "^(VERIFIED|NONE)$",
            "title": "Collector Reference Status"
          },
          "collector_reference_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Collector Reference Count"
          },
          "set_id": {
            "type": "string",
            "title": "Set Id"
          },
          "set_name": {
            "type": "string",
            "title": "Set Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "canonical_card_count",
          "collectible_printing_count",
          "collector_reference_status",
          "collector_reference_count",
          "set_id",
          "set_name"
        ],
        "title": "EditionDetail"
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "CARD_NOT_FOUND",
          "SET_NOT_FOUND",
          "EDITION_NOT_FOUND",
          "PRINTING_NOT_FOUND",
          "COLLECTOR_REFERENCE_NOT_FOUND",
          "KEYWORD_NOT_FOUND",
          "INVALID_FILTER",
          "INVALID_PAGINATION",
          "INVALID_SORT",
          "FORBIDDEN",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "description": "Known/expected error codes for the /v1 API.\n\nThis list is expected to grow in later phases as route handlers are\nimplemented (e.g. Phase 5+ will need to actually raise CARD_NOT_FOUND\nwhen a lookup misses) -- growing it is an additive change."
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeywordCatalogItem": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug",
            "example": "test-keyword"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Test Keyword"
          },
          "card_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Card Count",
            "example": 10
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "card_count"
        ],
        "title": "KeywordCatalogItem",
        "example": {
          "card_count": 10,
          "name": "Test Keyword",
          "slug": "test-keyword"
        }
      },
...
      "PaginatedCardsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CardSummary"
            },
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "PaginatedCardsResponse",
        "example": {
          "data": [],
          "pagination": {
            "has_next": false,
            "has_previous": false,
            "limit": 50,
            "page": 1,
            "pages": 0,
            "total": 0
          }
        }
      },
      "PaginatedEditionsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/EditionDetail"
            },
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "PaginatedEditionsResponse"
      },
      "PaginatedPrintingsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PrintingResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "PaginatedPrintingsResponse"
      },
...
      "PaginationMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "example": 50
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "example": 630
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages",
            "example": 13
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next",
            "example": true
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous",
            "example": false
          }
        },
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "pages",
          "has_next",
          "has_previous"
        ],
        "title": "PaginationMeta",
        "example": {
          "has_next": true,
          "has_previous": false,
          "limit": 50,
          "page": 1,
          "pages": 13,
          "total": 630
        }
      },
      "PrintingRarityCatalogItem": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Rare"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "example": "rare"
          },
          "printing_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Printing Count",
            "example": 84
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug",
          "printing_count"
        ],
        "title": "PrintingRarityCatalogItem",
        "example": {
          "name": "Rare",
          "printing_count": 84,
          "slug": "rare"
        }
      },
      "PrintingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "card_id": {
            "type": "string",
            "title": "Card Id"
          },
          "card_name": {
            "type": "string",
            "title": "Card Name"
          },
          "card_type": {
            "type": "string",
            "nullable": true,
            "title": "Card Type"
          },
          "set_id": {
            "type": "string",
            "title": "Set Id"
          },
          "set_name": {
            "type": "string",
            "title": "Set Name"
          },
          "edition_id": {
            "type": "string",
            "nullable": true,
            "title": "Edition Id"
          },
          "edition_name": {
            "type": "string",
            "nullable": true,
            "title": "Edition Name"
          },
          "collector_number": {
            "type": "string",
            "nullable": true,
            "title": "Collector Number"
          },
          "printed_number": {
            "type": "string",
            "title": "Printed Number"
          },
          "rarity": {
            "type": "string",
            "nullable": true,
            "title": "Rarity"
          },
          "collector_class": {
            "type": "string",
            "nullable": true,
            "title": "Collector Class"
          },
          "treatment": {
            "type": "string",
            "nullable": true,
            "title": "Treatment"
          },
          "finish": {
            "type": "string",
            "nullable": true,
            "title": "Finish"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "serial_numbered": {
            "type": "boolean",
            "title": "Serial Numbered"
          },
          "serial_total": {
            "type": "integer",
            "nullable": true,
            "title": "Serial Total"
          },
          "origin": {
            "type": "string",
            "pattern": "^(UPSTREAM|UPSTREAM_AND_CURATED|CURATED_REFERENCE)$",
            "title": "Origin"
          },
          "source_rarity": {
            "type": "string",
            "nullable": true,
            "title": "Source Rarity"
          },
          "source_variant": {
            "type": "string",
            "nullable": true,
            "title": "Source Variant"
          },
          "collector_reference_number": {
            "type": "string",
            "nullable": true,
            "title": "Collector Reference Number"
          },
          "collector_reference_card_name": {
            "type": "string",
            "nullable": true,
            "title": "Collector Reference Card Name"
          },
          "collector_reference_rarity": {
            "type": "string",
            "nullable": true,
            "title": "Collector Reference Rarity"
          },
          "collector_reference_variant": {
            "type": "string",
            "nullable": true,
            "title": "Collector Reference Variant"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/CardImageResponse"
            },
            "type": "array",
            "title": "Images"
          }
        },
        "type": "object",
        "required": [
          "id",
          "card_id",
          "card_name",
          "set_id",
          "set_name",
          "printed_number",
          "language",
          "serial_numbered",
          "origin"
        ],
        "title": "PrintingResponse"
      },
      "RarityCatalogItem": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Rare"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "example": "rare"
          },
          "card_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Card Count",
            "example": 42
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug",
          "card_count"
        ],
        "title": "RarityCatalogItem",
        "example": {
          "card_count": 42,
          "name": "Rare",
          "slug": "rare"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}